requirements.txt
amqp==2.6.1
asgiref==3.2.1
billiard==3.6.3.0
boto3==1.15.6
botocore==1.18.6
celery==4.4.7
certifi==2018.11.29
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
dj-database-url==0.5.0
Django==3.1.1
django-celery-results==1.2.1
django-configurations==2.2
django-crispy-forms==1.9.2
django-debug-toolbar==2.2
django-dynamic-formsets==0.0.8
django-storages==1.10.1
flake8==3.8.3
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
jmespath==0.10.0
kombu==4.6.11
mccabe==0.6.1
prompt-toolkit==3.0.7
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pyflakes==2.2.0
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.3
requests==2.24.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
vine==1.3.0
wcwidth==0.2.5
whitenoise==5.2.0
wincertstore==0.2
zipp==3.1.0
